Fix JWT security tests for deactivated-account 401s.
Missing customer rows now fail closed as 401 instead of authenticating
with no role. Align tests and disambiguate the account-status helper.

+++ backend/src/test/java/com/gpstore/security/CustomerAccountStatusServiceTest.java
@@ -51,7 +51,7 @@ class CustomerAccountStatusServiceTest {
 
     @Test
     void nullCustomerIdIsNotUsable() {
-        assertFalse(service.isUsable(null));
+        assertFalse(service.isUsable((Long) null));
     }
 
     @Test
